refactor(Matcher): Rename PredicateMatchers to BaseMatchers.
